refactor(canon): sépare UPgraders (ombrelle/société) de La Fusée (produit)
Bug d'identité : l'ADVE étiqueté « UPgraders » affichait « La Fusée » comme nom
de business — or UPgraders est la SOCIÉTÉ (qui vend) et La Fusée son PRODUIT
(l'Industry OS). Confusion explicitement interdite par la KB UPGRADERS-LAFUSEE.
Décision opérateur 2026-06-24 : deux ADVE distincts, UPgraders en marque ombrelle.

- `upgraders-canon.ts` (ancien) → renommé `lafusee-canon.ts` : c'est l'ADVE de
  La Fusée (nomMarque « La Fusée », PLATFORM). Exports → LAFUSEE_*. Contenu intact.
- `upgraders-canon.ts` (nouveau) : ADVE de la marque OMBRELLE UPgraders
  (nomMarque « UPgraders », SERVICE). Hérite (spread) du socle La Fusée
  contract-COMPLETE et surcharge les seuls champs d'identité :
    · A : nomMarque, description (société-fixer), secteur, brandNature SERVICE,
      noyauIdentitaire (l'opérateur qui possède/opère ses outils)
    · D : positionnement (maison-mère de La Fusée + Argos), lexique portfolio
    · V : produitsCatalogue = La Fusée (OS) + Argos (média) + conseil fixer + Guilde
  E/R/T/I/S hérités (ADN de marque partagé). Complétude garantie par construction
  (umbrella ⊇ La Fusée, vérifié : 32/17/25/23/9/14/16/20 clés, 0 surcharge vidée).
- `seed-upgraders.ts` : seede DEUX stratégies sous le client « UPgraders » via un
  helper partagé — l'ombrelle « UPgraders » + le produit « La Fusée — Industry OS ».

Importateurs inchangés : canon-sync / gauges / test HARD complétude / initiative
test consomment toujours UPGRADERS_* (= l'ombrelle désormais correcte). tsc +
eslint clean. Le test HARD de complétude reste vert (structure héritée).
